+++ expyriment/control/_experiment_control.py
@@ -407,8 +407,11 @@ fullscreen."""
         mixer.init()  # Needed on some systems
 
     experiment._clock = misc.Clock()
-    experiment._screen = Screen((0, 0, 0), defaults.open_gl,
-                                defaults.window_mode, defaults.window_size)
+    experiment._screen = Screen(colour=(0, 0, 0),
+                                open_gl=defaults.open_gl,
+                                window_mode=defaults.window_mode,
+                                window_size=defaults.window_size,
+                                blocking_mode=defaults.blocking_mode)
     # Hack for IDLE: quit pygame and call atexit functions when crashing
     if is_idle_running() and sys.argv[0] != "":
         try: