+++ expyriment/io/_files.py
@@ -625,18 +625,19 @@ class EventFile(OutputFile):
         try:
             display = repr(expyriment._active_exp.screen.window_size)
             window_mode = repr(expyriment._active_exp.screen.window_mode)
-            opengl = repr(expyriment._active_exp.screen.open_gl)
+            open_gl = repr(expyriment._active_exp.screen.open_gl)
+            blocking_mode = repr(expyriment._active_exp.screen.blocking_mode)
         except:
             display = "unknown"
             window_mode = "unknown"
-            opengl = "unknown"
+            open_gl = "unknown"
 
         self.write_comment("sha1: {0}".format(
                                     expyriment.get_experiment_secure_hash()))
         self.write_comment("modules: {0}".format(
                             expyriment._secure_hash.module_hashes_as_string()))
-        self.write_comment("display: size={0}, window_mode={1}, opengl={2}"
-                           .format(display, window_mode, opengl))
+        self.write_comment("display: size={0}, window_mode={1}, open_gl={2}," \
+" blocking_mode={3}".format(display, window_mode, open_gl, blocking_mode))
         self.write_comment("os: {0}".format(uname()))
 
         self.write_line("Time,Type,Event,Value,Detail,Detail2")