+++ CHANGES.md
@@ -41,6 +41,7 @@ New Features:
   max_width parameter has been defined
 - new feature: too long words in text boxes will be trimmed automatically, 
   this function can be switch off
+- control.defaults.blocking_mode for setting the blocking_mode
 
 Changed:
 - ParallelPort: Polling will now return 4 bits, including also the status of
@@ -48,6 +49,7 @@ Changed:
   (ATTENTION: Byte order in the output of other status lines has changed!)
 - changes at Simon example
 - the property Shape.size has been renamed to shape.shape_size
+- io.screen has new parameter to change the type of sync_screen blocking
 
 Fixed:
 - bug in io.TouchScreenButton crash if duration expired